pyproject.toml
[project]
name = "finance-tracker"
version = "0.1.0"
description = "Telegram bot for tracking personal finance"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "google-genai>=1.56.0",
    "pytest>=8.0.0",
]

